requirements.txt
pyjwt==2.3.0
fastapi==0.68.0
uvicorn==0.14.0
pydantic==1.8.2
python-multipart==0.0.5
tortoise-orm==0.17.8
bcrypt==3.2.0
passlib==1.7.4
pre-commit==2.15.0